+++ test/object-store/sync/client_reset.cpp
@@ -212,28 +212,27 @@ TEST_CASE("sync: client reset", "[client reset]") {
         size_t before_callback_invoctions = 0;
         size_t after_callback_invocations = 0;
         std::mutex mtx;
-        local_config.sync_config->notify_before_client_reset = [&](SharedRealm local, SharedRealm remote) {
+        local_config.sync_config->notify_before_client_reset = [&](SharedRealm before) {
             std::lock_guard<std::mutex> lock(mtx);
             ++before_callback_invoctions;
-            REQUIRE(local);
-            REQUIRE(local->is_frozen());
-            REQUIRE(local->read_group().get_table("class_object"));
-            REQUIRE(local->config().path == local_config.path);
-
-            REQUIRE(remote);
-            REQUIRE(remote->is_frozen());
-            REQUIRE(remote->read_group().get_table("class_object"));
-            REQUIRE(remote->config().path == fresh_path);
-
+            REQUIRE(before);
+            REQUIRE(before->is_frozen());
+            REQUIRE(before->read_group().get_table("class_object"));
+            REQUIRE(before->config().path == local_config.path);
             REQUIRE(util::File::exists(local_config.path));
-            REQUIRE(util::File::exists(fresh_path));
         };
-        local_config.sync_config->notify_after_client_reset = [&](SharedRealm local) {
+        local_config.sync_config->notify_after_client_reset = [&](SharedRealm before, SharedRealm after) {
             std::lock_guard<std::mutex> lock(mtx);
             ++after_callback_invocations;
-            REQUIRE(local);
-            REQUIRE(local->read_group().get_table("class_object"));
-            REQUIRE(local->config().path == local_config.path);
+            REQUIRE(before);
+            REQUIRE(before->is_frozen());
+            REQUIRE(before->read_group().get_table("class_object"));
+            REQUIRE(before->config().path == local_config.path);
+            REQUIRE(after);
+            REQUIRE(!after->is_frozen());
+            REQUIRE(after->read_group().get_table("class_object"));
+            REQUIRE(after->config().path == local_config.path);
+            REQUIRE(after->current_transaction_version() > before->current_transaction_version());
         };
 
         Results results;